- updated wallet connect view - fixed wallet topup scroll - fixed collectibles ui and hover
