+++ app/src/components/WalletAccount/AccountMenu/AccountMenu.vue
@@ -1,13 +1,5 @@
 <template>
   <v-card :flat="$vuetify.breakpoint.smAndDown" width="400" class="account-menu">
-    <v-dialog v-model="showQrScanner" :eager="true" :width="qrLoading ? 0 : 600" @click:outside="closeQRScanner">
-      <div class="qr-scan-container">
-        <QrcodeStream :camera="camera" :style="camera === 'off' && { display: 'none' }" @decode="onDecodeQr" @init="onInit" />
-        <v-btn class="close-btn" icon aria-label="Close QR Scanner" title="Close QR Scanner" @click="closeQRScanner">
-          <v-icon>$vuetify.icons.close</v-icon>
-        </v-btn>
-      </div>
-    </v-dialog>
     <v-list class="pb-0 mb-2">
       <v-list-item>
         <v-list-item-avatar class="ml-2 mr-3">
@@ -26,13 +18,8 @@
             <div>{{ t('accountMenu.account') }}</div>
           </div>
         </v-list-item-title>
-        <v-list-item-icon v-if="$vuetify.breakpoint.xsOnly && hasStreamApiSupport">
-          <div class="mr-5">
-            <v-btn small class="wallet-connect-btn" icon title="Capture QR" aria-label="Capture QR" @click="toggleWC">
-              <v-icon v-if="(wcConnectorSession && wcConnectorSession.connected) || false" size="16">$vuetify.icons.disconnect</v-icon>
-              <v-icon v-else size="16">$vuetify.icons.walletconnect</v-icon>
-            </v-btn>
-          </div>
+        <v-list-item-icon v-if="$vuetify.breakpoint.xsOnly">
+          <WalletConnect class="mr-5" />
         </v-list-item-icon>
       </v-list-item>
     </v-list>
@@ -135,15 +122,14 @@
 
 <script>
 import { BroadcastChannel } from 'broadcast-channel'
-import log from 'loglevel'
-import { QrcodeStream } from 'vue-qrcode-reader'
-import { mapActions, mapGetters, mapMutations, mapState } from 'vuex'
+import { mapActions, mapGetters, mapState } from 'vuex'
 
 import { DISCORD, GITHUB, TWITTER } from '../../../utils/enums'
 import { addressSlicer, broadcastChannelOptions, getEtherScanAddressLink, getUserEmail } from '../../../utils/utils'
 import ExportQrCode from '../../helpers/ExportQrCode'
 import LanguageSelector from '../../helpers/LanguageSelector'
 import ShowToolTip from '../../helpers/ShowToolTip'
+import WalletConnect from '../../helpers/WalletConnect'
 import AccountImport from '../AccountImport'
 
 export default {
@@ -152,7 +138,7 @@ export default {
     ExportQrCode,
     AccountImport,
     LanguageSelector,
-    QrcodeStream,
+    WalletConnect,
   },
   props: {
     headerItems: {
@@ -165,14 +151,11 @@ export default {
       accountImportDialog: false,
       DISCORD,
       camera: 'off',
-      qrErrorMsg: '',
-      showQrScanner: false,
-      qrLoading: true,
       hasStreamApiSupport: true,
     }
   },
   computed: {
-    ...mapState(['userInfo', 'selectedAddress', 'selectedCurrency', 'currencyData', 'networkType', 'wcConnectorSession']),
+    ...mapState(['userInfo', 'selectedAddress', 'selectedCurrency', 'currencyData', 'networkType']),
     ...mapGetters({
       wallets: 'walletBalances',
     }),
@@ -210,17 +193,8 @@ export default {
       return []
     },
   },
-  watch: {
-    qrErrorMsg(value) {
-      if (value) {
-        this.setErrorMsg(value)
-        this.qrErrorMsg = ''
-      }
-    },
-  },
   methods: {
     ...mapActions(['logOut', 'updateSelectedAddress', 'initWalletConnect', 'disconnectWalletConnect']),
-    ...mapMutations(['setErrorMsg']),
     etherscanAddressLink(address) {
       return getEtherScanAddressLink(address, this.networkType.host)
     },
@@ -247,58 +221,6 @@ export default {
         })
         selectedAddressChannel.close()
       }
-    },
-    toggleWC() {
-      if (this.wcConnectorSession?.connected) {
-        this.disconnectWalletConnect()
-      } else {
-        this.camera = 'auto'
-        this.showQrScanner = true
-      }
-    },
-    async onDecodeQr(result) {
-      try {
-        log.info(result, 'qr decoded')
-        await this.initWalletConnect({ uri: result })
-      } catch (error) {
-        log.error(error)
-      } finally {
-        this.camera = 'off'
-        this.showQrScanner = false
-      }
-    },
-    async onInit(promise) {
-      try {
-        await promise
-        this.qrLoading = false
-      } catch (error) {
-        log.error(error)
-        if (error.name === 'NotAllowedError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorNeedCameraPermission'
-          log.error('ERROR: you need to grant camera access permisson')
-        } else if (error.name === 'NotFoundError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorNoCamera'
-          log.error('ERROR: no camera on this device')
-        } else if (error.name === 'NotSupportedError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorSecureContextRequired'
-          log.error('ERROR: secure context required (HTTPS, localhost)')
-        } else if (error.name === 'NotReadableError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorCameraAlreadyInUse'
-          log.error('ERROR: is the camera already in use?')
-        } else if (error.name === 'OverconstrainedError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorInstalledCamerasAreNotSuitable'
-          log.error('ERROR: installed cameras are not suitable')
-        } else if (error.name === 'StreamApiNotSupportedError') {
-          this.qrErrorMsg = 'accountMenu.qrErrorStreamAPINotSupported'
-          log.error('ERROR: Stream Api not supported')
-
-          this.hasStreamApiSupport = false
-        }
-      }
-    },
-    closeQRScanner() {
-      this.camera = 'off'
-      this.showQrScanner = false
     },
   },
 }
