+++ app/src/components/WalletHome/CollectiblesList/CollectiblesList.scss
@@ -14,3 +14,11 @@
   font-size: 10px;
   line-height: 12px;
 }
+
+.asset-note {
+  position: absolute;
+  right: 16px;
+  bottom: 16px;
+  font-size: 10px;
+  line-height: 12px;
+}
