+++ app/src/components/helpers/WalletConnect/index.js
@@ -1,0 +1,1 @@
+export { default } from './WalletConnect.vue'
