+++ app/src/containers/WalletHistory/WalletHistory.scss
@@ -1,54 +1,16 @@
+.page-title {
+  font-size: 20px;
+  line-height: 23px;
+}
 .filter-selector {
-  background: #FFFFFF;
-  border: 1px solid #F5F5F5;
+  background: #ffffff;
+  border: 1px solid #f5f5f5;
   box-sizing: border-box;
   box-shadow: 0px 14px 28px rgba(46, 91, 255, 0.06);
   border-radius: 3px;
   &.theme--dark {
-    background: #2F3136;
+    background: #2f3136;
     border: 1px solid #343539;
     box-shadow: 0px 14px 28px rgba(92, 108, 127, 0.06);
   }
 }
-/* NEW UI */
-// .nav-selector.v-input.v-text-field--outlined {
-//   ::v-deep & > .v-input__control > .v-input__slot {
-//     box-shadow: none;
-//     border-radius: 0;
-//     background: none;
-//   }
-//   &.theme--dark {
-//     border: 1px solid #343539;
-//     ::v-deep & > .v-input__control > .v-input__slot {
-//       fieldset {
-//         border: 0
-//       }
-//     }
-//   }
-// }
-// ::v-deep .nav-selector {
-//   @include navSelector();
-//   border-radius: 3px;
-//   .v-select__selection {
-//     margin-left: 8px;
-//   }
-//   &.transaction {
-//     max-width: 190px;
-//   }
-//   &.period {
-//     max-width: 190px;
-//   }
-//   .v-icon {
-//     color: var(--v-text_2-base);
-//   }
-//   @media screen and (max-width: 600px) {
-//     &.transaction {
-//       max-width: 100%;
-//       width: 100%;
-//     }
-//     &.period {
-//       max-width: 100%;
-//       width: 100%;
-//     }
-//   }
-// }
