+++ app/src/containers/WalletHistory/WalletHistory.vue
@@ -2,7 +2,9 @@
   <v-container class="wallet-activity" :class="$vuetify.breakpoint.xsOnly ? 'px-4' : ''">
     <v-layout mt-3 wrap>
       <v-flex xs12 md7>
-        <div class="text_2--text font-weight-bold display-1 float-left">{{ t('walletActivity.transactionActivities') }}</div>
+        <div class="text_2--text font-weight-bold float-left page-title" :class="{ 'display-1': $vuetify.breakpoint.width > 390 }">
+          {{ t('walletActivity.transactionActivities') }}
+        </div>
       </v-flex>
       <v-flex xs12 md5 :class="$vuetify.breakpoint.xsOnly ? 'mt-7' : ''">
         <v-layout mx-n2>
