+++ app/src/containers/WalletHome/WalletHomeMain/WalletHomeMain.scss
@@ -1,3 +1,7 @@
+.page-title {
+  font-size: 20px;
+  line-height: 23px;
+}
 
 .card-total {
   .display-4 {
