+++ app/src/containers/WalletSettings/WalletSettings.scss
@@ -1,7 +1,10 @@
-
 @import '../../scss/elevation-one.mixin';
 
 .wallet-settings {
+  .page-title {
+    font-size: 20px;
+    line-height: 23px;
+  }
   .v-expansion-panels {
     .v-expansion-panel-header {
       border-top: 0;
@@ -11,7 +14,7 @@
     }
     .v-expansion-panel {
       &::before {
-        @include elevationOne()
+        @include elevationOne();
       }
       &:not(:first-child)::after {
         border: 0;
