+++ app/src/containers/WalletSettings/WalletSettings.vue
@@ -2,7 +2,9 @@
   <v-container class="wallet-settings" :class="$vuetify.breakpoint.xsOnly ? 'px-4' : ''">
     <v-layout wrap align-start :class="$vuetify.breakpoint.xsOnly ? 'mt-2' : 'mt-3'">
       <v-flex xs6>
-        <div class="font-weight-bold text_1--text display-1 float-left">{{ t('walletSettings.settings') }}</div>
+        <div class="font-weight-bold text_1--text float-left page-title" :class="{ 'display-1': $vuetify.breakpoint.width > 390 }">
+          {{ t('walletSettings.settings') }}
+        </div>
       </v-flex>
       <v-flex xs6>
         <QuickAddress />
