+++ app/src/containers/WalletTopup/WalletTopupHome/WalletTopupHome.scss
@@ -1,0 +1,4 @@
+.page-title {
+  font-size: 20px;
+  line-height: 23px;
+}
