+++ app/src/containers/WalletTopup/WalletTopupHome/WalletTopupHome.vue
@@ -1,7 +1,7 @@
 <template>
   <v-container class="wallet-topup-view pt-6" :class="$vuetify.breakpoint.xsOnly ? 'px-4' : ''">
     <div class="d-flex align-center">
-      <div class="font-weight-bold display-1 text-left text_2--text">
+      <div class="font-weight-bold text-left text_2--text page-title" :class="{ 'display-1': $vuetify.breakpoint.width > 390 }">
         <span v-if="selectedProvider && !$vuetify.breakpoint.xsOnly">
           {{ t('walletTopUp.purchaseVia') }}
           <span class="text-capitalize">{{ selectedProvider }}</span>
@@ -75,3 +75,7 @@ export default {
   },
 }
 </script>
+
+<style lang="scss" scoped>
+@import 'WalletTopupHome.scss';
+</style>
