+++ app/src/containers/WalletTransfer/WalletTransfer.scss
@@ -1,4 +1,8 @@
 .wallet-transfer {
+  .page-title {
+    font-size: 20px;
+    line-height: 23px;
+  }
   .currency-rate {
     margin-bottom: 2px;
   }
